requirements.txt
# Taiwan CCTV Viewer Dependencies
# Pin versions for reproducible builds

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
starlette==0.35.1

# HTTP Client
httpx==0.26.0

# Image Processing
pillow==10.2.0
opencv-python==4.9.0.80
numpy<2.0

# Machine Learning / Detection
ultralytics==8.1.0
torch==2.1.2
torchvision==0.16.2

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
greenlet==3.0.3

# WebSocket
websockets==12.0

# Tracking
filterpy==1.4.5

# Configuration
pyyaml==6.0.1
pydantic>=2.7.0
pydantic-settings>=2.2.0

# Observability
prometheus-client==0.19.0
psutil==5.9.8

# Security & Authentication
certifi==2024.2.2

# External Integration
anduril-lattice-sdk==4.0.0

# Optional: Redis (for distributed caching)
# redis==5.0.1

# Development / Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx[http2]==0.26.0

# Linting / Formatting (optional)
# black==24.1.0
# flake8==7.0.0
# mypy==1.8.0
# isort==5.13.2
